+++ lib/common/delay.h
@@ -7,6 +7,6 @@ typedef void(*interval_cb_func) (void);
 
 void os_delay(uint32_t ms);
 void sleep(uint32_t ms);
-void usleep(uint32_t us);
+void usleep(uint16_t us);
 
 #endif
